chore: rename Props to PropsList and remove StorybookLink
