+++ packages/f36-docs-utils/src/PropsTable/PropertyValue/PropertyValue.tsx
