+++ packages/f36-docs-utils/src/index.ts
@@ -1,1 +1,1 @@
-export * from './Props';
+export * from './PropsTable';
